split blockdiagonalmat
