+++ SofaKernel/modules/Sofa.Compat/src/SofaBaseLinearSolver/DiagonalMatrix.h
@@ -24,6 +24,8 @@
 #include <sofa/linearalgebra/DiagonalMatrix.h>
 
 // SOFA_DEPRECATED_HEADER("v21.12", "v22.06", "sofa/linearalgebra/DiagonalMatrix.h")
+// BlockDiagonalMatrix was located in DiagonalMatrix previously
+#include <sofa/linearalgebra/BlockDiagonalMatrix.h>
 
 namespace sofa::component::linearsolver
 {
