+++ SofaKernel/modules/Sofa.LinearAlgebra/src/sofa/linearalgebra/BTDMatrix.h
@@ -115,6 +115,7 @@ public:
 
     typedef Bloc SubMatrixType;
     typedef sofa::type::Mat<N,N,Real> BlocType;
+    typedef BlocFullMatrix<N, T> InvMatrixType;
     // return the dimension of submatrices when requesting a given size
     static Index getSubMatrixDim(Index) { return BSIZE; }
